backend/requirements.txt
# Django Core
Django>=5.2,<5.3
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
django-filter>=24.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.2.0

# Redis & Caching
redis>=5.0.0
django-redis>=5.4.0
hiredis>=3.0.0

# Celery
celery>=5.4.0
django-celery-beat>=2.8.0
django-celery-results>=2.5.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
google-auth>=2.27.0
requests>=2.31.0

# Security
argon2-cffi>=23.1.0
python-decouple
cryptography>=42.0.0

# URL & QR
validators>=0.22.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
nanoid>=2.0.0

reportlab==4.4.9

# Geo IP
geoip2>=4.8.0
user-agents>=2.2.0

# DNS Verification
dnspython>=2.6.0

# AWS S3
boto3>=1.34.0
django-storages>=1.14.0

# Stripe
stripe>=10.0.0

# Email - SendGrid
sendgrid>=6.11.0
django-sendgrid-v5>=1.2.0

# Monitoring & Logging
sentry-sdk[django,celery]>=2.0.0
python-json-logger>=2.0.0

# API Documentation
drf-spectacular>=0.28.0
drf-spectacular-sidecar>=2024.7.1

# CSV/Excel Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
factory-boy>=3.3.0
faker>=30.0.0
responses>=0.25.0
freezegun>=1.4.0

# Dev Tools
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
django-stubs>=4.2.0

# Production Server
gunicorn>=22.0.0
uvicorn[standard]>=0.30.0
whitenoise>=6.7.0

# Health Checks
django-health-check>=3.18.0
